requirements.txt
# Core Flask Framework
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.2
Flask-Migrate==4.1.0

# Database Migration
alembic==1.16.4

# Forms and Validation
WTForms==3.2.1
email-validator==2.2.0

# Google AI
google-generativeai==0.8.5

# PDF Processing
pypdf==5.7.0

# Location Services
geopy==2.4.1

# HTTP Requests
requests==2.32.4

# Environment Variables
python-dotenv==1.1.1

# Security
bcrypt==4.3.0

# WSGI Server for Production
gunicorn==21.2.0

# Essential Dependencies
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
itsdangerous==2.2.0
blinker==1.9.0
blinker==1.9.0
